feat: Enhance xtelebot functionality and improve test coverage
- Add new methods to xtelebot for handling inline queries
- Implement error handling for API requests in x/x.go
- Expand test cases in x_test.go and xtelebot_test.go
- Update types in xtelebot/types.go for better API compatibility
- Refactor main example in xtelebot_example/main.go for clarity

+++ xtelebot/xtelebot_test.go
@@ -55,7 +55,7 @@ func TestGetMe(t *testing.T) {
 	user, err := bot.GetMe(context.Background())
 	assert.NoError(t, err)
 	assert.NotNil(t, user)
-	assert.Equal(t, 123, user.ID)
+	assert.Equal(t, int64(123), user.ID)
 	assert.True(t, user.IsBot)
 	assert.Equal(t, "TestBot", user.FirstName)
 	assert.Equal(t, "test_bot", user.UserName)
